Minor Update to Water Script
by Nathan McClain

+++ Assets/Scripts/PlayerInventory_JG.cs
@@ -6,14 +6,15 @@ public class PlayerInventory_JG : MonoBehaviour {
 	public bool VineSeed;
 	public bool TreeSeed;
 
-	public float Water = 10f; 
+	public int Water = 10;
+	//Display Water Variable
 
-	public void TakeAway (float amount)
+	public void TakeAway (int amount)
 	{
 		// Decrement the player's water.
 		Water-= 1;
 	}
-	public void Add (float amount)
+	public void Add (int amount)
 	{
 		// Add to the player's water.
 		Water+= 1;
